Create m.file
In MATLAB, you can generate the sound of a major second interval using functions and tools for audio signal processing. This can be done using functions like audiowrite to save the audio and functions like chirp to generate the sound of a major second interval. However, generating an accurate major second interval depends on factors such as the exact frequency and desired duration. Below is a simple example code for generating the sound of a major second interval in MATLAB:
In this code, fs represents the sampling frequency of the audio signal, which is set to 44100 Hz here. duration specifies the duration of the generated sound in seconds. f calculates the frequency of the major second interval. Finally, sin(2*pi*f*t) generates a sound wave with the specified frequency, and sound(y, fs) plays the sound.